requirements.txt
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphene-django==3.0.2
psycopg2==2.9.6
python-decouple==3.8
whitenoise==6.4.0